fix: address absence assertion review feedback
